callback/requirements.txt
Flask==3.1.2
gunicorn==23.0.0
redis==7.1.0
flasgger==0.9.7.1
minio==7.2.8
langchain-core==0.3.72
reportlab==4.4.2
python-docx==1.2.0